Conservative governor tweaks for effectiveness
Originally from the Vibrant/Epic Kernel. Ported to the G2 for battery life improvements at no loss of performance at the moment.

These conservative governor tweaks are strictly for performance and responsiveness.  Currently in Beta testing right now.  Original commit is in the Epic Kernel.  As of now this leads to no loss of battery life.